Create mergeValidations function to merge the results of validations
Specifically it respects the `suppresses` mechanism.
